+++ scopeql-parser/CHANGELOG.md
@@ -6,7 +6,7 @@ All significant changes to the ScopeQL Parser be documented in this file.
 
 ### Breaking Changes
 
-* Drop "command" feature flag and related tokens (`BackSlack` and `CANCEL`).
+* Drop "command" feature flag and related tokens (`BackSlash` and `CANCEL`).
 
 ## v0.2.0 (2026-02-13)